+++ app/src/dayjs.ts
@@ -1,6 +1,6 @@
 import dayjs from "dayjs";
-import dayjs_relativeTime from "dayjs/plugin/relativeTime";
+import dayjs_weekOfYear from "dayjs/plugin/weekOfYear";
 
-dayjs.extend(dayjs_relativeTime);
+dayjs.extend(dayjs_weekOfYear);
 
 export default dayjs;